Add Node.js MySQL demo and generalize mysqlsniff script
The Node.js MySQL demo is very similar to the Java version, called
Databasey. It reads from MySQL and stumbles upon an intentionally
slow stored procedure invocation. The mysqlsniff script has been
updated to allow tracing the client side of this invocation by
attaching to the libpthread:__write probe. Unfortunately, the user-
space call stack for the Node.js process doesn't show the originator
of the MySQL query, but rather a stack deep inside the module. This
is due to Node's async nature.