java: Add skeleton for two contention monitoring exercises
1) JVMTI agent that tracks MonitorContendedEnter{,ed}
2) BCC tool that tracks monitor__contended__enter{,ed} USDT probes

The latter approach requires -XX:+ExtendedDTraceProbes, but there's
a lot more infrastructure already in place, and it doesn't require
an injection into the target. We can also track call stacks fairly
easily, and we get the class name as a USDT probe argument, so no
need to reach into JVM structures for that.